fix: file URI reformatting in DISLoader
Inherited problem from the time when Corelyzer was Win and Mac only: DISLoader checked for Mac OS, if positive it did nothing to the file path, if negative it changed the file path formatting for Windows, i.e. it also changed file formatting to Windows on a Linux system. After this fix, DISLoader will check for Windows OS and change the path if positive, if negative it will do nothing to the file path. Thus, it is not only a fix for Linux but should work with all POSIX systems.