fix: select only active section by default
(2nd attempt, first commit reverted) The trim dialogue was preset on "selected and deeper sections". This is not safe, because the user expects to work with the selection. Thus, "selected section only" should be the default and "selected and deeper sections" an option that is actively chosen by the user. This fix changes the order that "selected section only" is default.